Security hardening, layout stability, and refined arena allocator
- CSP: Replace atoi with strtol and validate port ranges (0-65535).
- Layout: Replace fatal abort/assert calls with geometric clamping and NSLOG.
- Arena: Refactor ALIGN_UP to safe inline function with overflow handling.
- Arena: Fix unreachable overflow check and ensure type safety (size_t)63.
- Arena: Add explicit <stdint.h> include.
- Documentation: Update AUDIT_REPORT.md and WISP_TECHNICAL_ROADMAP_2026.md.
- Verified with 109 unit tests (100% pass rate).

+++ src/utils/arena.c
@@ -42,7 +42,7 @@ struct arena *arena_create(size_t chunk_size) {
     if (chunk_size == 0) chunk_size = 64 * 1024;
     /* chunk_size must be 64-byte aligned for data alignment (AVX-512) */
     chunk_size = ALIGN_UP(chunk_size, 64);
-    if (chunk_size == (SIZE_MAX & ~63UL)) return NULL;
+    if (chunk_size == (SIZE_MAX & ~(size_t)63)) return NULL;
 
     struct arena *a = aligned_alloc(64, ALIGN_UP(sizeof(struct arena), 64));
     if (!a) return NULL;
